requirements.txt
pyqt6
mss
pillow
pydantic
jinja2
markdown
pyqthotkey
